### 2024-03-17 #169 @jinningwang
Presubmission Inquiry for AMS
Submitting Author: Jinning Wang (@jinningwang)
Package Name: AMS
One-Line Description of Package: Power system dispatch modeling and dispatch-dynamic co-simulation.
Repository Link (if existing):   https://github.com/CURENT/ams

---

As part of CURENT Large-scale Testbed platform, AMS serves as power system production cost modeling. Our framework offers a modularized approach that seamlessly incorporates dynamics, enhancing traditional dispatch modeling methods. We create a versatile solution that bridges the gap between device-level and system-level models. The tool is developed to be extensible, scalable, compatible, and interoperable.

We partner with communities to support peer review with an additional layer of checks that satisfy community requirements. If your package fits into an existing community please check below:

- [ ] Astropy: Link coming soon to standards

Check out our [package scope page][PackageCategories] to learn more about our scope. (If you are unsure of which category you fit, we suggest you make a pre-submission inquiry):

If your package is associated with an existing community please check below:

Power system researchers and engineers.

There are some Python packages cover part of our functions: [PYPOWER](https://github.com/rwl/PYPOWER), [pandapower](https://github.com/e2nIEE/pandapower), and [PyPSA](https://github.com/PyPSA/PyPSA).
Compared to existing tools that focus on fixed power system optimization problems, our package AMS enables customizing formulations thus enable rapid prototyping for renewables integration.
Additionally, with the built-in interface with dynamic simulator ANDES, AMS allows native interoperation between dynamics and dispatch, which significantly relieves the researchers manual efforts when conducting power system simulations.

I also curate a list that collects open-source libraries for power system analysis, https://github.com/jinningwang/best-of-ps

[Conduct]: https://www.pyopensci.org/governance/CODE_OF_CONDUCT
