### 2024-03-23 #169 @isabelizimm
Hello there! Thank you so much for this submission, it’s obvious there has been a lot of work put into AMS! I’ve taken a look at AMS, and think it would be in scope, but have a few thoughts.
I _think_ the best category for this would be *data munging*:

> Data munging tools transform data in a way that makes further analysis possible (as [defined on Wikipedia](https://en.wikipedia.org/wiki/Data_wrangling)). [...] This category focuses on tools for handling data in specific formats that scientists may be interested in working with.

My impression is that power system researchers can use AMS to read in files and solve/update simulations, and load models to ANDES.

Are you able able to give some feedback on what you think of this category, that I have understood this package's functionality?
