### 2024-03-23 #169 @jinningwang
@isabelizimm Thanks for your response!

Your impression to AMS is mostly correct and let me further clarify it.

> My impression is that power system researchers can use AMS to read in files and solve/update simulations,

Yes.

> and load models to ANDES.

I'm not sure if I understand this correctly. I would say AMS can convert AMS case file to ANDES case file, and thus ANDES is capable of another type of power system simulation.

Overall, I feel it's reasonable to classify AMS to scope Data processing/munging.

Additionally, in the future, is it helpful to have a Domain Specific "power system analysis"? I notice that there are more and more open-source tools developed for power system researchers.
