### 2024-03-23 01cd62c
Fix formulation of line angle diff constr
