+++ ams/system.py
@@ -405,14 +405,13 @@ class System(andes_System):
         adjust_rate = []
         default_rate = 999
         for rate in [self.Line.rate_a, self.Line.rate_b, self.Line.rate_c]:
-            if np.all(rate.v == 0):
+            if np.any(rate.v == 0):
                 adjust_rate.append(rate.name)
                 rate.v[rate.v == 0] = default_rate
         if adjust_rate:
             adjusted_rate = ', '.join(adjust_rate)
             msg = f"Zero line rates detacted in {adjusted_rate}, "
             msg += f"adjusted to {default_rate}."
-            msg += "\nIf expect a line outage, please set 'u' to 0."
             logger.info(msg)
         # === no device addition or removal after this point ===
         self.calc_pu_coeff()   # calculate parameters in system per units
