### 2024-03-24 e90a0f1
Add more common params in Group
