+++ ams/shared.py
@@ -6,9 +6,11 @@ This module is supplementary to the ``andes.shared`` module.
 import logging
 from functools import wraps
 from datetime import datetime
+from collections import OrderedDict
 
 import cvxpy as cp
 
+from andes.shared import pd
 
 logger = logging.getLogger(__name__)
 
@@ -36,3 +38,51 @@ def require_MIP_solver(f):
         return f(*args, **kwargs)
 
     return wrapper
+
+
+ppc_cols = OrderedDict([
+    ('bus', ['bus_i', 'type', 'pd', 'qd', 'gs', 'bs', 'area', 'vm', 'va',
+             'baseKV', 'zone', 'vmax', 'vmin', 'lam_p', 'lam_q',
+             'mu_vmax', 'mu_vmin']),
+    ('branch', ['fbus', 'tbus', 'r', 'x', 'b', 'rate_a', 'rateB_b', 'rate_c',
+                'ratio', 'angle', 'status', 'angmin',
+                'angmax', 'pf', 'qf', 'pt', 'qt', 'mu_sf', 'mu_st',
+                'mu_angmin', 'mu_angmax']),
+    ('gen', ['bus', 'pg', 'qg', 'qmax', 'qmin', 'vg', 'mbase', 'status',
+             'pmax', 'pmin', 'pc1', 'pc2', 'qc1min', 'qc1max', 'qc2min',
+             'qc2max', 'ramp_agc', 'ramp_10', 'ramp_30', 'ramp_q',
+             'apf', 'mu_pmax', 'mu_pmin', 'mu_qmax', 'mu_qmin']),
+    ('gencost', ['model', 'startup', 'shutdown', 'n', 'c', 'coeffs'])
+])
+
+
+def ppc2df(ppc, model='bus', ppc_cols=ppc_cols):
+    """
+    Convert PYPOWER dict to pandas DataFrame.
+
+    Parameters
+    ----------
+    ppc : dict
+        PYPOWER case dict.
+    model : str
+        Model name.
+    ppc_cols : dict
+        Column names.
+
+    Returns
+    -------
+    pandas.DataFrame
+        DataFrame.
+
+    Examples
+    --------
+    >>> import ams
+    >>> sp = ams.system.example()
+    >>> ppc = ams.io.pypower.system2ppc(sp)
+    >>> ppc_bus = ams.shared.ppc2df(ppc, 'bus')
+    """
+    if model not in ppc_cols.keys():
+        raise ValueError(f"Invalid model {model}")
+
+    df = pd.DataFrame(ppc[model], columns=ppc_cols[model][0:ppc[model].shape[1]])
+    return df
