### 2024-03-25 744c974
Add more common_params in group
