### 2024-03-25 0762924
Fix init sequence issue
