### 2024-03-25 0a065b7
Fix ppc gencost cols
