### 2024-03-25 b56e030
Fix sparse matrix type in matprocessor
