+++ ams/routines/dcopf.py
@@ -5,7 +5,7 @@ import logging
 
 import numpy as np
 from ams.core.param import RParam
-from ams.core.service import NumOp, NumOpDual
+from ams.core.service import NumOp, NumOpDual, VarSelect
 
 from ams.routines.routine import RoutineBase
 
@@ -75,6 +75,11 @@ class DCOPF(RoutineBase):
                           name='p0', tex_name=r'p_{g, 0}',
                           unit='p.u.',
                           model='StaticGen', src='pg0')
+        # --- bus ---
+        self.buss = RParam(info='Bus slack',
+                           name='buss', tex_name=r'B_{us,s}',
+                           model='Slack', src='bus',
+                           no_parse=True,)
         # --- load ---
         self.pd = RParam(info='active demand',
                          name='pd', tex_name=r'p_{d}',
@@ -151,6 +156,13 @@ class DCOPF(RoutineBase):
                         unit='rad',
                         name='aBus', tex_name=r'\theta_{bus}',
                         model='Bus', src='a',)
+        self.csb = VarSelect(info='select slack bus',
+                             name='csb', tex_name=r'c_{sb}',
+                             u=self.aBus, indexer='buss',
+                             no_parse=True,)
+        self.sba = Constraint(info='align slack bus angle',
+                              name='sbus', is_eq=True,
+                              e_str='csb@aBus',)
         self.pi = Var(info='nodal price',
                       name='pi', tex_name=r'\pi',
                       unit='$/p.u.',
