+++ ams/io/matpower.py
@@ -38,6 +38,9 @@ def mpc2system(mpc: dict, system) -> bool:
 
     Compared to the original one, this function includes the generator cost data.
 
+    Note that `mbase` in mpc is converted to `Sn`, but it is not actually used in
+    MATPOWER nor AMS.
+
     Parameters
     ----------
     system : andes.system.System
@@ -90,11 +93,10 @@ def mpc2system(mpc: dict, system) -> bool:
     if mpc['gen'].shape[1] <= 10:  # missing data
         mpc_gen = np.zeros((mpc['gen'].shape[0], 21), dtype=np.float64)
         mpc_gen[:, :10] = mpc['gen']
-        mbase = mpc['gen'][:, 6]
-        mpc_gen[:, 16] = system.PV.Ragc.default * mbase / 60
-        mpc_gen[:, 17] = system.PV.R10.default * mbase
-        mpc_gen[:, 18] = system.PV.R30.default * mbase
-        mpc_gen[:, 19] = system.PV.Rq.default * mbase / 60
+        mpc_gen[:, 16] = system.PV.Ragc.default * base_mva / 60
+        mpc_gen[:, 17] = system.PV.R10.default * base_mva
+        mpc_gen[:, 18] = system.PV.R30.default * base_mva
+        mpc_gen[:, 19] = system.PV.Rq.default * base_mva / 60
     else:
         mpc_gen = mpc['gen']
     for data in mpc_gen:
@@ -104,28 +106,26 @@ def mpc2system(mpc: dict, system) -> bool:
         # 10   11   12      13      14      15      16        17
         # ramp_30  ramp_q  apf
         # 18       19      20
-
         bus_idx = int(data[0])
         gen_idx += 1
         vg = data[5]
         status = int(data[7])
-        mbase = data[6]
-        pg = data[1] / mbase
-        qg = data[2] / mbase
-        qmax = data[3] / mbase
-        qmin = data[4] / mbase
-        pmax = data[8] / mbase
-        pmin = data[9] / mbase
-        pc1 = data[10] / mbase
-        pc2 = data[11] / mbase
-        qc1min = data[12] / mbase
-        qc1max = data[13] / mbase
-        qc2min = data[14] / mbase
-        qc2max = data[15] / mbase
-        ramp_agc = 60 * data[16] / mbase  # from MW/min to MW/h
-        ramp_10 = data[17] / mbase  # from MW to MW/h
-        ramp_30 = data[18] / mbase  # from MW to MW/h
-        ramp_q = 60 * data[19] / mbase  # from MVAr/min to MVAr/h
+        pg = data[1] / base_mva
+        qg = data[2] / base_mva
+        qmax = data[3] / base_mva
+        qmin = data[4] / base_mva
+        pmax = data[8] / base_mva
+        pmin = data[9] / base_mva
+        pc1 = data[10] / base_mva
+        pc2 = data[11] / base_mva
+        qc1min = data[12] / base_mva
+        qc1max = data[13] / base_mva
+        qc2min = data[14] / base_mva
+        qc2max = data[15] / base_mva
+        ramp_agc = 60 * data[16] / base_mva  # MW/min -> MW/h
+        ramp_10 = data[17] / base_mva  # MW -> MW/h
+        ramp_30 = data[18] / base_mva  # MW -> MW/h
+        ramp_q = 60 * data[19] / base_mva  # MVAr/min -> MVAr/h
         apf = data[20]
 
         uid = system.Bus.idx2uid(bus_idx)
[…]
