### 2024-03-26 2a7bc56
Ignore demonstration notebooks in CI
