+++ .github/workflows/pythonapp.yml
@@ -39,7 +39,7 @@ jobs:
     - shell: bash -el {0}
       name: Test notebooks.
       run: |
-        pytest --nbmake examples --ignore=examples/demonstration
+        pytest --nbmake examples --ignore=examples/verification
     - name: Build a distribution if tagged
       if: github.event_name == 'push' && startsWith(github.ref, 'refs/tags') && matrix.os == 'ubuntu-latest' && matrix.python-version == '3.9'
       run: |
