### 2024-04-04 0cf32db
Add a loss factor in dc2ac
