+++ ams/routines/rted.py
@@ -181,11 +181,14 @@ class RTED(DCOPF, RTEDBase, SFRBase):
         cost += f'+ t dot sum({_to_sum})'
         self.obj.e_str = cost
 
-    def dc2ac(self, **kwargs):
+    def dc2ac(self, kloss=1.1, **kwargs):
         """
         Convert the RTED results with ACOPF.
 
-        Overload ``dc2ac`` method.
+        Parameters
+        ----------
+        kloss : float, optional
+            The loss factor for the conversion. Defaults to 1.2.
         """
         if self.exec_time == 0 or self.exit_code != 0:
             logger.warning('RTED is not executed successfully, quit conversion.')
@@ -196,14 +199,25 @@ class RTED(DCOPF, RTEDBase, SFRBase):
         pmax0 = self.system.StaticGen.get(src='pmax', attr='v', idx=pr_idx)
         p00 = self.system.StaticGen.get(src='p0', attr='v', idx=pr_idx)
 
-        # solve ACOPF
+        # --- ACOPF ---
+        # scale up load
+        pq_idx = self.system.StaticLoad.get_idx()
+        pd0 = self.system.StaticLoad.get(src='p0', attr='v', idx=pq_idx).copy()
+        qd0 = self.system.StaticLoad.get(src='q0', attr='v', idx=pq_idx).copy()
+        self.system.StaticLoad.set(src='p0', attr='v', idx=pq_idx, value=pd0 * kloss)
+        self.system.StaticLoad.set(src='q0', attr='v', idx=pq_idx, value=qd0 * kloss)
+        # preserve generator reserve
         ACOPF = self.system.ACOPF
         pmin = pmin0 + self.prd.v
         pmax = pmax0 - self.pru.v
         self.system.StaticGen.set(src='pmin', attr='v', idx=pr_idx, value=pmin)
         self.system.StaticGen.set(src='pmax', attr='v', idx=pr_idx, value=pmax)
         self.system.StaticGen.set(src='p0', attr='v', idx=pr_idx, value=self.pg.v)
+        # run ACOPF
         ACOPF.run()
+        # scale load back
+        self.system.StaticLoad.set(src='p0', attr='v', idx=pq_idx, value=pd0)
+        self.system.StaticLoad.set(src='q0', attr='v', idx=pq_idx, value=qd0)
         if not ACOPF.exit_code == 0:
             logger.warning('<ACOPF> did not converge, conversion failed.')
             # NOTE: mock results to fit interface with ANDES
