### 2024-04-06 9548d33
Improve verify_pf messaging
