### 2024-04-06 a981f85
Add test on verify_pf
