### 2024-04-19 448b497
Add functions to build system matrices
