+++ ams/models/line.py
@@ -1,6 +1,6 @@
 from andes.models.line.line import LineData
 from andes.core.param import NumParam
-from andes.shared import deg2rad
+from andes.shared import deg2rad, np, spmatrix
 
 from ams.core.model import Model
 
@@ -11,6 +11,9 @@ class Line(LineData, Model):
 
     The model is also used for two-winding transformer. Transformers can set the
     tap ratio in ``tap`` and/or phase shift angle ``phi``.
+
+    Note that the bus admittance matrix is built on fly and is not stored in the
+    object.
 
     Notes
     -----
@@ -39,3 +42,160 @@ class Line(LineData, Model):
         self.rate_a.default = 999.0
         self.rate_b.default = 999.0
         self.rate_c.default = 999.0
+
+        # NOTE: following parameters are prepared for building matrices
+        # they are initialized here but populated in ``System.setup()``.
+        self.a1a = None
+        self.a2a = None
+
+    def build_y(self):
+        """
+        Build bus admittance matrix. Copied from ``andes.models.line.line.Line``.
+
+        Returns
+        -------
+        Y : spmatrix
+            Bus admittance matrix.
+        """
+
+        nb = self.system.Bus.n
+
+        y1 = self.u.v * (self.g1.v + self.b1.v * 1j)
+        y2 = self.u.v * (self.g2.v + self.b2.v * 1j)
+        y12 = self.u.v / (self.r.v + self.x.v * 1j)
+        m = self.tap.v * np.exp(1j * self.phi.v)
+        m2 = self.tap.v**2
+        mconj = np.conj(m)
+
+        # build self and mutual admittances into Y
+        Y = spmatrix((y12 + y1 / m2), self.a1a, self.a1a, (nb, nb), 'z')
+        Y -= spmatrix(y12 / mconj, self.a1a, self.a2a, (nb, nb), 'z')
+        Y -= spmatrix(y12 / m, self.a2a, self.a1a, (nb, nb), 'z')
+        Y += spmatrix(y12 + y2, self.a2a, self.a2a, (nb, nb), 'z')
+
+        return Y
+
+    def build_Bp(self, method='fdpf'):
+        """
+        Function for building B' matrix.
+
+        Parameters
+        ----------
+        method : str
+            Method for building B' matrix. Choose from 'fdpf', 'fdbx', 'dcpf'.
+
+        Returns
+        -------
+        Bp : spmatrix
+            B' matrix.
+        """
+        nb = self.system.Bus.n
+
+        if method not in ("fdpf", "fdbx", "dcpf"):
+            raise ValueError(f"Invalid method {method}; choose from 'fdpf', 'fdbx', 'dcpf'")
+
+        # Build B prime matrix -- FDPF
+        # `y1`` neglects line charging shunt, and g1 is usually 0 in HV lines
+        # `y2`` neglects line charging shunt, and g2 is usually 0 in HV lines
+        y1 = self.u.v * self.g1.v
+        y2 = self.u.v * self.g2.v
+
+        # `m` neglected tap ratio
+        m = np.exp(self.phi.v * 1j)
+        mconj = np.conj(m)
+        m2 = np.ones(self.n)
+
+        if method in ('fdxb', 'dcpf'):
+            # neglect line resistance in Bp in XB method
+            y12 = self.u.v / (self.x.v * 1j)
+        else:
+            y12 = self.u.v / (self.r.v + self.x.v * 1j)
+
[…]
