### 2024-04-19 b2f794a
Add test for matrices building
