### 2024-04-20 8347fff
Add group common parameters
