+++ ams/opt/omodel.py
@@ -746,9 +746,9 @@ class OModel:
         self.constrs = OrderedDict()
         self.obj = None
         self.initialized = False
-        self._parsed = False
+        self.parsed = False
 
-    def _parse(self, no_code=True):
+    def parse(self, no_code=True):
         """
         Parse the optimization model from the symbolic description.
 
@@ -815,8 +815,8 @@ class OModel:
             else:
                 logger.warning(f"{rtn.class_name} has no objective function!")
                 _, s = elapsed(t0)
-                self._parsed = False
-                return self._parsed
+                self.parsed = False
+                return self.parsed
         _, s = elapsed(t0)
         logger.debug(f"Parse Objective in {s}")
 
@@ -827,8 +827,8 @@ class OModel:
         _, s = elapsed(t0)
         logger.debug(f"Parse Expressions in {s}")
 
-        self._parsed = True
-        return self._parsed
+        self.parsed = True
+        return self.parsed
 
     def init(self, no_code=True):
         """
@@ -850,8 +850,8 @@ class OModel:
         """
         t_setup, _ = elapsed()
 
-        if not self._parsed:
-            self._parse(no_code=no_code)
+        if not self.parsed:
+            self.parse(no_code=no_code)
 
         if self.rtn.type == 'PF':
             _, s_setup = elapsed(t_setup)
