### 2024-04-20 ea564c8
Format Routine.get return type to be consistent with input idx
