### 2024-04-20 0e1ea2e
Add test on routine.get return type
