### 2024-04-20 89e0a60
Add test on symbol generation
