### 2024-04-20 48cde51
Add test on gen trip and line trip
