### 2024-04-20 a0f6d5e
Add test on constraint and objective values
