### 2024-04-20 58ad10f
Rename MatProcessor make as build, refactor _makeC into four subfunctions
