+++ ams/core/matprocessor.py
@@ -129,9 +129,6 @@ class MatProcessor:
         self.Cg = MParam(name='Cg', tex_name=r'C_g',
                          info='Generator connectivity matrix',
                          v=None, sparse=True)
-        self.Cs = MParam(name='Cs', tex_name=r'C_s',
-                         info='Slack connectivity matrix',
-                         v=None, sparse=True)
         self.Cl = MParam(name='Cl', tex_name=r'Cl',
                          info='Load connectivity matrix',
                          v=None, sparse=True)
@@ -139,14 +136,19 @@ class MatProcessor:
                           info='Shunt connectivity matrix',
                           v=None, sparse=True)
 
-    def make(self):
+    def build(self):
         """
-        Make the system matrices.
-
-        Note that this method will update all matrices in the class.
+        Build the system matrices.
+        It includes connectivity matrices: Cg, Cl, Csh, Cft, and CftT,
+        and bus matrices: Bbus, Bf, Pbusinj, and Pfinj.
         """
         t_mat, _ = elapsed()
-        self._makeC()
+        # --- connectivity matrices ---
+        _ = self.build_cg()
+        _ = self.build_cl()
+        _ = self.build_csh()
+        _ = self.build_cft()
+        # --- bus matrices ---
         self._makeBdc()
         _, s_mat = elapsed(t_mat)
         logger.debug(f"Built system matrices in {s_mat}.")
@@ -212,70 +214,115 @@ class MatProcessor:
         self.Bbus._v, self.Bf._v, self.Pbusinj._v, self.Pfinj._v = Bbus, Bf, Pbusinj, Pfinj
         return True
 
-    def _makeC(self):
+    def build_cg(self):
         """
-        Make connectivity matrix.
+        Build generator connectivity matrix Cg, and store it in the MParam `Cg`.
+
+        Returns
+        -------
+        Cg : spmatrix
+            Generator connectivity matrix.
         """
         system = self.system
 
         # common variables
         nb = system.Bus.n
         ng = system.StaticGen.n
-        nl = system.Line.n
-        npq = system.PQ.n
-        nsh = system.Shunt.n
 
-        # --- Cg ---
         # bus indices: idx -> uid
         idx_gen = system.StaticGen.get_idx()
         u_gen = system.StaticGen.get(src='u', attr='v', idx=idx_gen)
         on_gen = np.flatnonzero(u_gen)  # uid of online generators
         on_gen_idx = [idx_gen[i] for i in on_gen]  # idx of online generators
-        on_gen_bus_idx = system.StaticGen.get(src='bus', attr='v', idx=on_gen_idx)
+        on_gen_bus = system.StaticGen.get(src='bus', attr='v', idx=on_gen_idx)
 
-        row = np.array([system.Bus.idx2uid(x) for x in on_gen_bus_idx])
+        row = np.array([system.Bus.idx2uid(x) for x in on_gen_bus])
         col = np.array([idx_gen.index(x) for x in on_gen_idx])
-        Cg = c_sparse((np.ones(len(on_gen_idx)), (row, col)), (nb, ng))
+        self.Cg._v = c_sparse((np.ones(len(on_gen_idx)), (row, col)), (nb, ng))
+        return self.Cg._v
 
-        # --- Cl ---
+    def build_cl(self):
+        """
+        Build load connectivity matrix Cl, and store it in the MParam `Cl`.
+
+        Returns
+        -------
+        Cl : spmatrix
+            Load connectivity matrix.
+        """
+        system = self.system
+
[…]
