### 2024-04-20 c359205
Refactor test_service to have separate tests on MatProcessor
