### 2024-04-20 6a07d3f
Add tests on connectivity matrices
