### 2024-04-21 f57200e
Refactor _makeBdc into separate methods
