### 2024-04-21 0bf1b7b
Add note in build_ptdf docstring
