### 2024-04-21 b146ac6
Update build_ptdf docstring
