### 2024-04-21 12846d4
Use safe_div
