### 2024-04-21 6a18035
Add test on LODF
