### 2024-04-21 016f3b5
Add test on build_otdf
