+++ tests/test_mats.py
@@ -157,4 +157,27 @@ class TestMatProcessor(unittest.TestCase):
             dplf = plf1 - plf0
             dplfc = -lodf[:, oline] * dplf[oline]
 
-            np.testing.assert_allclose(dplf, dplfc, atol=1e-1)
+            np.testing.assert_allclose(dplf, dplfc, atol=1e-7)
+
+    def test_otdf(self):
+        """
+        Test `OTDF`.
+        """
+
+        for case in self.cases:
+            ss = ams.load(ams.get_case(case),
+                          setup=True, default_config=True, no_output=True)
+            # build matrices
+            ss.mats.build()
+
+            oline_idx = ss.Line.idx.v[1]
+
+            otdf = ss.mats.build_otdf(line=oline_idx)
+
+            ss.Line.set(src='u', attr='v', idx=oline_idx, value=0)
+            ss.DCPF.run()
+
+            plf = ss.DCPF.plf.v
+            plfc = otdf@(ss.mats.Cg._v@ss.DCPF.pg.v - ss.mats.Cl._v@ss.DCPF.pd.v)
+
+            np.testing.assert_allclose(plf, plfc, atol=1e-7)
