### 2024-04-21 fd79831
Support on Routine.get(idx=pd.Series)
