### 2024-04-21 83753cd
Refactor tesst_mats to combine connectivity matrices test
